requirements.txt
streamlit
opencv-python
pytesseract
layoutparser
transformers
numpy
torch
pillow
pandas
plotly
gradio